:truck: | Move some files + light refactor
