+++ components/GetDate.tsx
@@ -1,0 +1,8 @@
+export function getDate(): string {
+  return new Date().toLocaleDateString(undefined, {
+    weekday: "long",
+    year: "numeric",
+    month: "long",
+    day: "numeric",
+  });
+}
